refactor auth into public pkg
